requirements.txt
# ── Core ──────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
pydantic[email]>=2.6.0

# ── Database & Integration ────────────────────────
supabase>=2.4.0
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
requests>=2.31.0

# ── Utilities & Security ──────────────────────────
python-multipart>=0.0.9
Pillow>=10.2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# ── AI & Machine Learning ──────────────────────────
tensorflow>=2.15.0
